@@ -1,6 +1,6 @@
 module github.com/dronestock/drone
 
-go 1.17
+go 1.18
 
 require (
 	github.com/goexl/gex v0.2.6
@@ -27,7 +27,7 @@ require (
 	go.uber.org/multierr v1.8.0 // indirect
 	go.uber.org/zap v1.23.0 // indirect
 	golang.org/x/crypto v0.0.0-20220829220503-c86fa9a7ed90 // indirect
-	golang.org/x/sys v0.0.0-20220829200755-d48e67d00261 // indirect
+	golang.org/x/sys v0.0.0-20220907062415-87db552b00fd // indirect
 	golang.org/x/text v0.3.7 // indirect
 )
 
